+++ supertokens_python/recipe/emailpassword/api/password_reset.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     )
 
 from supertokens_python.exceptions import raise_bad_input_exception
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 from .utils import validate_form_fields_or_throw_error
 
@@ -47,7 +47,9 @@ async def handle_password_reset_api(
         raise_bad_input_exception("The password reset token must be a string")
 
     token = body["token"]
+    user_context = await default_user_context(api_options.request)
+
     response = await api_implementation.password_reset_post(
-        form_fields, token, api_options, {}
+        form_fields, token, api_options, user_context
     )
     return send_200_response(response.to_json(), api_options.response)