+++ supertokens_python/recipe/emailpassword/api/signin.py
@@ -22,7 +22,7 @@ if TYPE_CHECKING:
     )
 
 from supertokens_python.exceptions import raise_bad_input_exception
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 from .utils import validate_form_fields_or_throw_error
 
@@ -37,6 +37,10 @@ async def handle_sign_in_api(api_implementation: APIInterface, api_options: APIO
     form_fields = await validate_form_fields_or_throw_error(
         api_options.config.sign_in_feature.form_fields, form_fields_raw
     )
-    response = await api_implementation.sign_in_post(form_fields, api_options, {})
+    user_context = await default_user_context(api_options.request)
+
+    response = await api_implementation.sign_in_post(
+        form_fields, api_options, user_context
+    )
 
     return send_200_response(response.to_json(), api_options.response)