+++ supertokens_python/recipe/emailverification/api/generate_email_verify_token.py
@@ -17,7 +17,7 @@ from supertokens_python.recipe.emailverification.interfaces import (
     APIInterface,
     APIOptions,
 )
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 
 async def handle_generate_email_verify_token_api(
@@ -25,5 +25,9 @@ async def handle_generate_email_verify_token_api(
 ):
     if api_implementation.disable_generate_email_verify_token_post:
         return None
-    result = await api_implementation.generate_email_verify_token_post(api_options, {})
+    user_context = await default_user_context(api_options.request)
+
+    result = await api_implementation.generate_email_verify_token_post(
+        api_options, user_context
+    )
     return send_200_response(result.to_json(), api_options.response)