+++ supertokens_python/recipe/session/api/signout.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from supertokens_python.recipe.session.interfaces import APIInterface, APIOptions
 
-from supertokens_python.utils import send_200_response
+from supertokens_python.utils import default_user_context, send_200_response
 
 
 async def handle_signout_api(api_implementation: APIInterface, api_options: APIOptions):
@@ -27,7 +27,9 @@ async def handle_signout_api(api_implementation: APIInterface, api_options: APIO
         or api_implementation.signout_post is None
     ):
         return None
-    response = await api_implementation.signout_post(api_options, {})
+    user_context = await default_user_context(api_options.request)
+
+    response = await api_implementation.signout_post(api_options, user_context)
     if api_options.response is None:
         raise Exception("Should never come here")
     return send_200_response(response.to_json(), api_options.response)