+++ tests/__init__.py
@@ -11,3 +11,7 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
+
+import nest_asyncio  # type: ignore
+
+nest_asyncio.apply()  # type: ignore